refactored SearchBarResults into separate component
